v0.2.0
Finished implementing regex parser and CLI.

Pass the text to parse as the first arg on the command line, and the
regex word as the second.

If this process is successful, the word searched will be printed to
stdout. If it is not, an error message will appear in the same place.